requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
absl-py=2.0.0=pypi_0
accelerate=0.30.1=pypi_0
aiohttp=3.9.1=pypi_0
aiosignal=1.3.1=pypi_0
alembic=1.13.0=pypi_0
annotated-types=0.6.0=pypi_0
annoy=1.17.3=pypi_0
antlr4-python3-runtime=4.9.3=pypi_0
anyio=3.7.1=pypi_0
asttokens=2.4.1=pyhd8ed1ab_0
astunparse=1.6.3=pypi_0
async-timeout=4.0.3=pypi_0
attrs=23.1.0=pypi_0
beautifulsoup4=4.12.2=pypi_0
beir=2.0.0=pypi_0
bert-score=0.3.13=pypi_0
binutils_impl_linux-64=2.38=h2a08ee3_1
bitarray=2.9.2=pypi_0
blinker=1.7.0=pypi_0
blis=0.7.11=pypi_0
boilerpy3=1.0.7=pypi_0
bzip2=1.0.8=h7b6447c_0
ca-certificates=2024.2.2=hbcca054_0
cachetools=5.3.2=pypi_0
catalogue=2.0.10=pypi_0
certifi=2023.7.22=pypi_0
cffi=1.16.0=pypi_0
charset-normalizer=3.3.0=pypi_0
click=8.1.7=pypi_0
cloudpathlib=0.16.0=pypi_0
cloudpickle=3.0.0=pypi_0
colbert-ir=0.2.14=pypi_0
comcrawl=1.0.2=pypi_0
comm=0.1.4=pyhd8ed1ab_0
confection=0.1.3=pypi_0
construct=2.5.3=pypi_0
contourpy=1.2.0=pypi_0
cryptography=42.0.4=pypi_0
cuda-cudart=11.8.89=0
cudatoolkit=11.3.1=hb98b00a_13
cycler=0.12.1=pypi_0
cymem=2.0.8=pypi_0
databricks-cli=0.18.0=pypi_0
datasets=2.15.0=pypi_0
debugpy=1.6.7=py310h6a678d5_0
decorator=5.1.1=pyhd8ed1ab_0
dill=0.3.7=pypi_0
distro=1.9.0=pypi_0
docker=6.1.3=pypi_0
docker-pycreds=0.4.0=pypi_0
docopt=0.6.2=pypi_0
docstring-parser=0.16=pypi_0
dpr=1.0.0=pypi_0
editdistance=0.6.2=pypi_0
elastic-transport=8.10.0=pypi_0
elasticsearch=7.9.1=pypi_0
en-core-web-sm=3.7.1=pypi_0
entrypoints=0.4=pyhd8ed1ab_0
et-xmlfile=1.1.0=pypi_0
evaluate=0.4.1=pypi_0
exceptiongroup=1.1.3=pypi_0
executing=2.0.1=pyhd8ed1ab_0
faiss-gpu=1.7.2=pypi_0
fastapi=0.104.1=pypi_0
filelock=3.12.4=pypi_0
fitlog=0.9.15=pypi_0
flask=3.0.0=pypi_0
flatbuffers=23.5.26=pypi_0
fonttools=4.47.0=pypi_0
frozenlist=1.4.0=pypi_0
fsspec=2023.9.2=pypi_0
gast=0.5.4=pypi_0
gcc=12.1.0=h9ea6d83_10
gcc_impl_linux-64=12.1.0=hea43390_17
gdown=4.7.1=pypi_0
gevent=24.2.1=pypi_0
git-filter-repo=2.38.0=pypi_0
git-python=1.0.3=pypi_0
gitdb=4.0.11=pypi_0
gitpython=3.1.40=pypi_0
google-api-core=2.17.1=pypi_0
google-api-python-client=2.121.0=pypi_0
google-auth=2.25.1=pypi_0
google-auth-httplib2=0.2.0=pypi_0
google-auth-oauthlib=1.1.0=pypi_0
google-pasta=0.2.0=pypi_0
googleapis-common-protos=1.62.0=pypi_0
googlesearch-python=1.2.3=pypi_0
gputil=1.4.0=pypi_0
greenlet=3.0.2=pypi_0
grequests=0.7.0=pypi_0
grpcio=1.59.3=pypi_0
gunicorn=21.2.0=pypi_0
h11=0.14.0=pypi_0
h5py=3.10.0=pypi_0
haystack=0.42=pypi_0
httpcore=1.0.2=pypi_0
httplib2=0.22.0=pypi_0
httpx=0.26.0=pypi_0
huggingface-hub=0.19.4=pypi_0
hydra-core=1.3.2=pypi_0
icu=73.2=h59595ed_0
idna=3.4=pypi_0
importlib-metadata=7.0.0=pypi_0
intel-openmp=2023.1.0=hdb19cb5_46306
ipykernel=6.26.0=pyhf8b6a83_0
ipython=8.18.1=pyh707e725_3
itsdangerous=2.1.2=pypi_0
jedi=0.19.1=pyhd8ed1ab_0
jinja2=3.1.2=pypi_0
joblib=1.3.2=pypi_0
jsonlines=4.0.0=pypi_0
jupyter_client=7.3.4=pyhd8ed1ab_0
jupyter_core=5.5.0=py310hff52083_0
keras=2.15.0=pypi_0
kernel-headers_linux-64=2.6.32=he073ed8_16
kiwisolver=1.4.5=pypi_0
krippendorff=0.6.1=pypi_0
langcodes=3.3.0=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libblas=3.9.0=22_linux64_openblas
libcblas=3.9.0=22_linux64_openblas
libclang=16.0.6=pypi_0
libcublas=11.11.3.6=0
libfaiss=1.8.0=h5aaf3ed_0_cuda11.4.4
libffi=3.4.4=h6a678d5_0
libgcc=7.2.0=h69d50b8_2
libgcc-devel_linux-64=12.1.0=h1ec3361_17
libgcc-ng=13.2.0=h807b86a_3
libgfortran-ng=13.2.0=h69a702a_5
libgfortran5=13.2.0=ha4646dd_5
libgomp=13.2.0=h807b86a_3
libhwloc=2.9.1=hd6dc26d_0
liblapack=3.9.0=22_linux64_openblas
libopenblas=0.3.27=pthreads_h413a1c8_0
libsanitizer=12.1.0=ha89aaad_17
libsodium=1.0.18=h36c2ea0_1
libstdcxx-ng=13.2.0=h7e041cc_3
libuuid=1.41.5=h5eee18b_0
libxml2=2.10.4=hf1b16e4_1
mako=1.3.0=pypi_0
markdown=3.5.1=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.3=pypi_0
matplotlib=3.8.2=pypi_0
matplotlib-inline=0.1.6=pyhd8ed1ab_0
mdurl=0.1.2=pypi_0
memory-profiler=0.61.0=pypi_0
mkl=2023.1.0=h213fc3f_46344
ml-dtypes=0.2.0=pypi_0
mlflow=2.9.2=pypi_0
mpmath=1.3.0=pypi_0
multidict=6.0.4=pypi_0
multiprocess=0.70.15=pypi_0
murmurhash=1.0.10=pypi_0
ncurses=6.4=h6a678d5_0
ndg-httpsclient=0.5.1=pypi_0
nest-asyncio=1.5.8=pyhd8ed1ab_0
networkx=3.1=pypi_0
ninja=1.11.1.1=pypi_0
nltk=3.8.1=pypi_0
numpy=1.26.0=pypi_0
nvidia-cublas-cu12=12.1.3.1=pypi_0
nvidia-cuda-cupti-cu12=12.1.105=pypi_0
nvidia-cuda-nvrtc-cu12=12.1.105=pypi_0
nvidia-cuda-runtime-cu12=12.1.105=pypi_0
nvidia-cudnn-cu12=8.9.2.26=pypi_0
nvidia-cufft-cu12=11.0.2.54=pypi_0
nvidia-curand-cu12=10.3.2.106=pypi_0
nvidia-cusolver-cu12=11.4.5.107=pypi_0
nvidia-cusparse-cu12=12.1.0.106=pypi_0
nvidia-nccl-cu12=2.18.1=pypi_0
nvidia-nvjitlink-cu12=12.2.140=pypi_0
nvidia-nvtx-cu12=12.1.105=pypi_0
oauthlib=3.2.2=pypi_0
omegaconf=2.3.0=pypi_0
openai=1.27.0=pypi_0
openpyxl=3.1.2=pypi_0
openssl=3.2.1=hd590300_1
opt-einsum=3.3.0=pypi_0
packaging=23.2=pyhd8ed1ab_0
pandas=2.1.3=pypi_0
parso=0.8.3=pyhd8ed1ab_0
patsy=0.5.6=pypi_0
pefile=2023.2.7=pypi_0
peft=0.11.1=pypi_0
pexpect=4.8.0=pyh1a96a4e_2
pickleshare=0.7.5=py_1003
pillow=10.1.0=pypi_0
pip=23.2.1=py310h06a4308_0
platformdirs=4.1.0=pyhd8ed1ab_0
preshed=3.0.9=pypi_0
prompt-toolkit=3.0.41=pyha770c72_0
protobuf=4.23.4=pypi_0
psutil=5.9.0=py310h5eee18b_0
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
pyarrow=14.0.1=pypi_0
pyarrow-hotfix=0.6=pypi_0
pyasn1=0.5.1=pypi_0
pyasn1-modules=0.3.0=pypi_0
pycorenlp=0.3.0=pypi_0
pycparser=2.21=pypi_0
pydantic=2.4.2=pypi_0
pydantic-core=2.10.1=pypi_0
pygments=2.17.2=pyhd8ed1ab_0
pyjwt=2.8.0=pypi_0
pyopenssl=24.0.0=pypi_0
pyparsing=3.1.1=pypi_0
pysocks=1.7.1=pypi_0
python=3.10.13=h955ad1f_0
python-dateutil=2.8.2=pyhd8ed1ab_0
python-dotenv=1.0.0=pypi_0
python-ptrace=0.9.8=pypi_0
python_abi=3.10=2_cp310
pytrec-eval=0.5=pypi_0
pytz=2023.3.post1=pypi_0
pyyaml=6.0.1=pypi_0
pyzmq=25.1.0=py310h6a678d5_0
querystring-parser=1.2.4=pypi_0
rank-bm25=0.2.2=pypi_0
readline=8.2=h5eee18b_0
regex=2023.10.3=pypi_0
requests=2.31.0=pypi_0
requests-oauthlib=1.3.1=pypi_0
responses=0.18.0=pypi_0
rich=13.7.1=pypi_0
rouge-score=0.1.2=pypi_0
rsa=4.9=pypi_0
safetensors=0.4.0=pypi_0
scikit-learn=1.3.2=pypi_0
scipy=1.11.3=pypi_0
sentence-transformers=2.2.2=pypi_0
sentencepiece=0.1.99=pypi_0
sentry-sdk=2.1.1=pypi_0
setproctitle=1.3.3=pypi_0
setuptools=58.2.0=pypi_0
shtab=1.7.1=pypi_0
six=1.16.0=pyh6c4a22f_0
smart-open=6.4.0=pypi_0
smmap=5.0.1=pypi_0
sniffio=1.3.0=pypi_0
soundfile=0.12.1=pypi_0
soupsieve=2.5=pypi_0
spacy=3.7.2=pypi_0
spacy-legacy=3.0.12=pypi_0
spacy-loggers=1.0.5=pypi_0
sqlalchemy=2.0.23=pypi_0
sqlite=3.41.2=h5eee18b_0
sqlparse=0.4.4=pypi_0
srsly=2.4.8=pypi_0
stack_data=0.6.2=pyhd8ed1ab_0
starlette=0.27.0=pypi_0
statsmodels=0.14.1=pypi_0
swig=4.2.1=pypi_0
sympy=1.12=pypi_0
sysroot_linux-64=2.12=he073ed8_16
tabulate=0.9.0=pypi_0
tbb=2021.9.0=hf52228f_0
tensorboard=2.15.1=pypi_0
tensorboard-data-server=0.7.2=pypi_0
tensorflow=2.15.0.post1=pypi_0
tensorflow-estimator=2.15.0=pypi_0
tensorflow-io-gcs-filesystem=0.34.0=pypi_0
termcolor=2.4.0=pypi_0
thinc=8.2.1=pypi_0
threadpoolctl=3.2.0=pypi_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.15.0=pypi_0
torch=1.13.1+cu116=pypi_0
torchaudio=0.13.1+rocm5.2=pypi_0
torchvision=0.14.1+cu116=pypi_0
tornado=6.1=py310h5764c6d_3
tqdm=4.66.1=pypi_0
traitlets=5.14.0=pyhd8ed1ab_0
transformers=4.35.2=pypi_0
triton=2.1.0=pypi_0
trl=0.9.4=pypi_0
typer=0.9.0=pypi_0
typing_extensions=4.8.0=pyha770c72_0
tyro=0.8.4=pypi_0
tzdata=2023.3=pypi_0
ujson=5.8.0=pypi_0
uritemplate=4.1.1=pypi_0
urllib3=2.0.6=pypi_0
wandb=0.17.0=pypi_0
wasabi=1.1.2=pypi_0
wcwidth=0.2.12=pyhd8ed1ab_0
weasel=0.3.3=pypi_0
websocket-client=1.7.0=pypi_0
werkzeug=3.0.1=pypi_0
wget=3.2=pypi_0
wheel=0.41.2=py310h06a4308_0
wrapt=1.14.1=pypi_0
xxhash=3.4.1=pypi_0
xz=5.4.2=h5eee18b_0
yarl=1.9.4=pypi_0
zeromq=4.3.4=h2531618_0
zipp=3.17.0=pypi_0
zlib=1.2.13=h5eee18b_0
zope-event=5.0=pypi_0
zope-interface=6.2=pypi_0
